final draft

+++ company-service/src/main/java/com/ankitjain/company/service/service/CompanyService.java
@@ -29,7 +29,7 @@ public class CompanyService {
     public CompanyVO getCompanyDetail(Long companyCode) throws Exception {
         Optional<Company> company = this.companyRepository.findById(companyCode);
         if (company.isPresent()) {
-            ResponseEntity<Stock> response = this.restTemplate.exchange("http://STOCK-SERVICE/api/v1.0/stock/getLatest/" + companyCode, HttpMethod.GET, null, Stock.class);
+            ResponseEntity<Stock> response = this.restTemplate.exchange("http://STOCK-SERVICE/api/v1.0/market/stock/getLatest/" + companyCode, HttpMethod.GET, null, Stock.class);
             return CompanyVO.builder().company(company.get()).latestStockPrice(response.getBody().getStockPrice()).build();
         } else {
             throw new Exception("Company code not found");
@@ -40,17 +40,22 @@ public class CompanyService {
         List<CompanyVO> resList = new ArrayList<>();
         List<Company> companies = this.companyRepository.findAll();
         companies.forEach(company -> {
-            ResponseEntity<Stock> response = this.restTemplate.exchange("http://STOCK-SERVICE/api/v1.0/stock/getLatest/" + company.getCompanyCode(), HttpMethod.GET, null, Stock.class);
+            ResponseEntity<Stock> response = this.restTemplate.exchange("http://STOCK-SERVICE/api/v1.0/market/stock/getLatest/" + company.getCompanyCode(), HttpMethod.GET, null, Stock.class);
             resList.add(CompanyVO.builder()
-            .company(company)
-            .latestStockPrice(response.getBody().getStockPrice())
-            .build());
+                    .company(company)
+                    .latestStockPrice(response.getBody().getStockPrice())
+                    .build());
         });
         return resList;
     }
 
-    public void deleteCompanyWithAllStockPrices(Long companyCode) {
-        this.companyRepository.deleteById(companyCode);
-        this.restTemplate.exchange("http://STOCK-SERVICE/api/v1.0/stock/deleteAll/" + companyCode, HttpMethod.DELETE, null, Stock.class);
+    public void deleteCompanyWithAllStockPrices(Long companyCode) throws Exception {
+        Optional<Company> company = this.companyRepository.findById(companyCode);
+        if (company.isPresent()) {
+            this.companyRepository.deleteById(companyCode);
+            this.restTemplate.exchange("http://STOCK-SERVICE/api/v1.0/market/stock/deleteAll/" + companyCode, HttpMethod.DELETE, null, String.class);
+        } else {
+            throw new Exception("Company does not exist");
+        }
     }
 }
